Added more changed to Home Page.html
That file needs to have it’s name changed to something with no caps or
spaces. As well, I changed the images so they look a bit better. The
website is finished but we could polish it by adding a custom CSS file.
Aside from that, the website simply needs to link to the game and
that’s it for the website. As for the game, we’re pressed for time so
I’ll be mainly working on the side-panel beside the pong window. If I
can add images or any of our stretch goals i’ll try and get that done
as well.