Finalized (hopefully) and reorganized sound folder
Though what we have right now is absolutely fantastic and more than
what we need for the test i’m still not done with our game yet. I want
the other groups to look up to what we’ve done on Thursday’s class and
see us as the pinnacle of team-based software development. Because we
are.